ctcs2: Fix bug#793914: ctcs2 opens log file with O_SYNC
Commit is based on a patch by Klaus Wagner.

This commit disables O_SYNC logging by default.

It also adds two new tcf keywords 'synclog' and
'asynclog' that could be used to turn on and off
O_SYNC logging for individual testcases.

I've also added support for SYNC_LOGGING environment
variable that overrides the 'synclog' and 'asynclog'
commands and, if set, all logging is done with O_SYNC
wich is exactly the behavior we had previously.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>