ctcs2: Fix bug #797695
This changes the signal that is send to the process
group when the test is being killed from SIGINT
to SIGTERM (as the SIGINT is not delivered to the
child processes running in the background).

Also the loop that tries to kill the childrens
in the runtest shell script is removed (as this
only duplicated the code in the ctcs2).

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>