CTCS2: Fix terminal output for timeouted tests.
If test has run less than expected iterations and no internal error
happened it has timeouted.

Also change the test results to more readable and slightly longer
variants.

Note that the results are still wrong in the summary. Where tests that
do more than one iteration appears as successfull even when they
timeouted in subsequent runs. To fix this, we need to accumulate number
of expected iterations and propagate it in the results.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>